% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clusterEvalStay}
\alias{clusterEvalStay}
\title{Appends a NULL to the end of a brace so that the result stays on the workers and is not returned to the manager}
\usage{
clusterEvalStay(cl, expr)
}
\description{
Appends a NULL to the end of a brace so that the result stays on the workers and is not returned to the manager
}
