% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_csv_data.R
\name{gen_csv_data}
\alias{gen_csv_data}
\title{Generate synthetic CSV data}
\usage{
gen_csv_data(p, n, fname = default_csv_file(), MB = 100,
  sampler = function(n) signif(runif(n)), group_probs = rep(0.1, 10),
  prepend_group_column = TRUE, column_names = paste0("col", seq(p)),
  group_column_name = "g", ...)
}
\arguments{
\item{p}{number of columns}

\item{n}{number of rows. Specify only one of n or MB.}

\item{fname}{name of the file to save}

\item{MB}{size of the data in MB. An integer group column will make this a little smaller.}

\item{sampler}{function that creates random observations}

\item{prepend_group_column}{whether to include a}

\item{...}{arguments to \code{data.table:::fwrite}}
}
\value{
d generated data that was written to fname
}
\description{
Generate synthetic CSV data
}
