% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_groupby.R
\name{gen_data_groupby}
\alias{gen_data_groupby}
\title{Generate and save data for shuffled data group by experiment}
\usage{
gen_data_groupby(nfiles = 10L, ngroups = 8L, block_two = TRUE,
  block_magnitude = 1, rand_gen = runif,
  dir = file.path(default_data_dir(),
  sprintf("groupby_\%ifiles_\%igroups", nfiles, ngroups)), ...)
}
\arguments{
\item{nfiles}{number of files}

\item{ngroups}{number of distinct groups in the grouping column}

\item{block_two}{logical, whether to add two diagonal blocks to the distributions of the files and groups, which means that some groups are more likely to be located in particular files.}

\item{rand_gen}{random number generating function}

\item{dir}{directory to save files}

\item{...}{further arguments to gen_table}

\item{block_multiplier}{magnitude of block structure}
}
\description{
Generate and save data for shuffled data group by experiment
}
