% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_table.R
\name{gen_table}
\alias{gen_table}
\title{Generate synthetic tabular data}
\usage{
gen_table(n, p = 10, fname = default_csv_file(), MB = 100,
  sampler = function(n) signif(runif(n)), group_probs = rep(0.1, 10),
  prepend_group_column = TRUE, column_names = paste0("col", seq(p)),
  group_column_name = "g", writer = data.table::fwrite, ...)
}
\arguments{
\item{n}{number of rows. Specify only one of n or MB.}

\item{p}{number of columns}

\item{fname}{name of the file to save}

\item{MB}{size of the data in MB. An integer group column will make this a little smaller.}

\item{sampler}{function that creates random observations}

\item{prepend_group_column}{whether to include a}

\item{writer}{function to save generated data. Must have signature (data, fname, ...)}

\item{...}{arguments to writer}
}
\value{
d generated data that was written to fname
}
\description{
Generate synthetic tabular data
}
