% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_shuffle.R, R/two_groupby.R
\name{group_by_local_shuffle}
\alias{group_by_local_shuffle}
\alias{greedy_group_assign}
\alias{data_local_group_assign}
\title{Time GROUP BY operation}
\usage{
group_by_local_shuffle(dir, nworkers = 3L,
  assign_groups = data_local_group_assign, group_fun = median)

greedy_group_assign(P, nworkers)

data_local_group_assign(P, w)
}
\arguments{
\item{dir}{directory where data can be found}

\item{nworkers}{number of workers}

\item{assign_groups}{function to assign files and groups to workers, signature must match default}

\item{group_fun}{function to apply to each group}
}
\description{
Internally this uses data.table, because it's fast and we want the speed.
}
\section{Functions}{
\itemize{
\item \code{greedy_group_assign}: Greedy scheduling to assign groups in a GROUP BY to workers

\item \code{data_local_group_assign}: Data location aware scheduling to assign groups in a GROUP BY to workers
}}

