% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_by_in_memory.R
\name{group_by_data.table}
\alias{group_by_data.table}
\title{Time Data.table GROUP BY operation}
\usage{
group_by_data.table(fname = default_csv_file(), threads = 2L,
  by_col = "g", data_col = "col1", group_fun = median)
}
\arguments{
\item{fname}{name of the file to read}
}
\description{
Time Data.table GROUP BY operation
}
