% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_by_in_memory.R
\name{group_by_split_lapply}
\alias{group_by_split_lapply}
\title{Time GROUP BY operation}
\usage{
group_by_split_lapply(fname = default_csv_file(), by_col = "g",
  data_col = "col1", group_fun = median)
}
\arguments{
\item{fname}{name of the file to read}
}
\description{
Time GROUP BY operation
}
