\name{r_data_benchmarks-package}
\alias{r_data_benchmarks-package}
\alias{r_data_benchmarks}
\docType{package}
\title{
\packageTitle{r_data_benchmarks}
}
\description{
\packageDescription{r_data_benchmarks}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{r_data_benchmarks}
\packageIndices{r_data_benchmarks}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{r_data_benchmarks}

Maintainer: \packageMaintainer{r_data_benchmarks}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
